refactor: simplify repo
