--- lib/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  plugin: require('./postGraphilePlugin')
-}
